requirements.txt
botbuilder-core>=4.7.0
botbuilder-dialogs>=4.11.0
aiohttp